- Added a font settings button to preface and foreword screens. -  moved  Reading Settings Bottom Sheet out of the reader file and made it reusable for both reading screen and preface and foreword screens
