+++ lib/screens/foreword_preface/foreword_preface_screen.dart
@@ -4,7 +4,10 @@ import 'package:family_altar/screens/foreword_preface/bloc/foreword_preface_bloc
 import 'package:family_altar/theme/app_colors.dart';
 import 'package:family_altar/theme/app_fonts.dart';
 import 'package:family_altar/theme/app_icons.dart';
+import 'package:family_altar/theme/bloc/theme_bloc.dart';
+import 'package:family_altar/theme/bloc/theme_state.dart';
 import 'package:family_altar/utils/utilities.dart';
+import 'package:family_altar/widgets/reading_settings_bottom_sheet.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
@@ -104,6 +107,34 @@ class ForewordPrefaceScreen extends StatelessWidget {
                       size: AppIcons.getIconSize(IconSize.medium),
                     ),
                   ),
+                  IconButton(
+                    onPressed: () => showReadingSettingsBottomSheet(context),
+                    padding: const EdgeInsets.symmetric(horizontal: 8),
+                    constraints: const BoxConstraints(
+                      minWidth: 40,
+                      minHeight: 40,
+                    ),
+                    icon: Container(
+                      width: 28,
+                      height: 28,
+                      alignment: Alignment.center,
+                      decoration: BoxDecoration(
+                        shape: BoxShape.circle,
+                        border: Border.all(
+                          color: context.textColor,
+                          width: 1.5,
+                        ),
+                      ),
+                      child: Text(
+                        'A',
+                        style: AppFonts.bold(context).copyWith(
+                          fontSize: 14,
+                          color: context.textColor,
+                          height: 1,
+                        ),
+                      ),
+                    ),
+                  ),
                 ],
               ),
               body: GestureDetector(
@@ -188,31 +219,39 @@ class _ContentBody extends StatelessWidget {
     if (loadedState.section == Section.dailyReading) {
       return const SizedBox.shrink();
     }
-    return SingleChildScrollView(
-      controller: scrollController,
-      child: Scrollbar(
-        child: Center(
-          child: Padding(
-            padding: const EdgeInsets.all(16),
-            child: TweenAnimationBuilder<double>(
-              key: ValueKey(loadedState.section),
-              duration: const Duration(milliseconds: 200),
-              tween: Tween(begin: 0, end: 1),
-              builder: (context, value, child) {
-                return Opacity(opacity: value, child: child);
-              },
-              child: Text(
-                loadedState.page.text.replaceAll(
-                  RegExp(r'(?<!\n)\n(?!\n)'),
-                  ' ',
+    return BlocBuilder<ThemeBloc, ThemeState>(
+      builder: (context, themeState) {
+        final fontSize = themeState.readingFontSize;
+        return SingleChildScrollView(
+          controller: scrollController,
+          child: Scrollbar(
+            child: Center(
+              child: Padding(
+                padding: const EdgeInsets.all(16),
+                child: TweenAnimationBuilder<double>(
+                  key: ValueKey(loadedState.section),
+                  duration: const Duration(milliseconds: 200),
+                  tween: Tween(begin: 0, end: 1),
+                  builder: (context, value, child) {
+                    return Opacity(opacity: value, child: child);
+                  },
+                  child: Text(
+                    loadedState.page.text.replaceAll(
+                      RegExp(r'(?<!\n)\n(?!\n)'),
+                      ' ',
+                    ),
+                    textAlign: TextAlign.left,
+                    style: AppFonts.normal(
+                      context,
+                    ).copyWith(fontSize: fontSize, height: 1.2),
+                    textScaler: TextScaler.noScaling,
+                  ),
                 ),
-                textAlign: TextAlign.left,
-                style: AppFonts.normal(context),
               ),
             ),
           ),
-        ),
-      ),
+        );
+      },
     );
   }
 }
